pyproject.toml
[project]
name = "docbot"
version = "0.1.0"
description = "docbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "black",
    "tiktoken",
    "openai",
    "unstructured",
    "nltk",
    "fastapi",
    "jinja2",
    "uvicorn",
    "streamlit",
    "streamlit-chat",
    "tqdm",
    "isort",
    "langchainhub",
    "firecrawl-py",
    "langchain-community",
    "langsmith",
    "pillow",
    "python-dotenv",
    "langchain>=0.3.3",
    "langchain-openai>=0.0.5",
    "pydantic>=2.0.0",
    "azure-core",
    "langchain-ollama>=0.3.3",
    "langchain-core>=0.3.60",
    "langchain-community>=0.3.24",
    "tavily-python>=0.7.2",
    "flask>=3.1.1",
    "langsmith>=0.3.42",
    "blank>=0.1.0",
    "langchain-pinecone>=0.2.6",
    "pinecone",
    "pypdf>=5.5.0",
    "faiss-cpu>=1.11.0",
]
